Lazy import from the data submodule
The top-level import of data_dir in skimage/__init__.py accounted for
approximately 1/3 of the package import time. This PR removes this overhead.

This commit moves skimage/data/__init__.py to skimage/data/_fetchers.py
and then adds a new skimage/data/__init__.py that lazily loads from the
skimage.data submodules.
