+++ skimage/__init__.py
@@ -95,7 +95,11 @@ from ._shared.version_requirements import ensure_python_version
 ensure_python_version((3, 7))
 
 from ._shared import lazy
-__getattr__, __lazy_dir__, _ = lazy.attach(__name__, submodules)
+__getattr__, __lazy_dir__, _ = lazy.attach(
+    __name__,
+    submodules,
+    submod_attrs={'data': ['data_dir']}
+)
 
 
 def __dir__():
@@ -158,7 +162,6 @@ else:
                              img_as_ubyte,
                              img_as_bool,
                              dtype_limits)
-    from .data import data_dir
     from .util.lookfor import lookfor
 
 if 'dev' in __version__:
