+++ skimage/data/tests/test_data.py
@@ -1,6 +1,6 @@
 import numpy as np
 import skimage.data as data
-from skimage.data import image_fetcher
+from skimage.data._fetchers import image_fetcher
 from skimage import io
 from skimage._shared.testing import assert_equal, assert_almost_equal, fetch
 import os
